catch-up

+++ setup.py
@@ -1,8 +1,24 @@
 #!/usr/bin/env python2
 # -*- coding=utf-8 -*-
 
+from __future__ import print_function
+
 import sys
 import os
+
+try:
+    import xml.etree.ElementTree
+    print("Using xml.etree.ElementTree for XML processing")
+except ImportError as e:
+    sys.stderr.write(str(e) + "\n")
+    try:
+        import elementtree.ElementTree
+        print("Using elementtree.ElementTree for XML processing")
+    except ImportError as e:
+        sys.stderr.write(str(e) + "\n")
+        sys.stderr.write("Please install ElementTree module from\n")
+        sys.stderr.write("http://effbot.org/zone/element-index.htm\n")
+        sys.exit(1)
 
 from setuptools import setup
 
@@ -13,33 +29,19 @@ if float("%d.%d" % sys.version_info[:2]) < 2.6:
     sys.stderr.write("S3cmd requires Python 2.6 or newer.\n")
     sys.exit(1)
 
-try:
-    import xml.etree.ElementTree as ET
-    print "Using xml.etree.ElementTree for XML processing"
-except ImportError, e:
-    sys.stderr.write(str(e) + "\n")
+## Remove 'MANIFEST' file to force
+## distutils to recreate it.
+## Only in "sdist" stage. Otherwise
+## it makes life difficult to packagers.
+if len(sys.argv) > 1 and sys.argv[1] == "sdist":
     try:
-        import elementtree.ElementTree as ET
-        print "Using elementtree.ElementTree for XML processing"
-    except ImportError, e:
-        sys.stderr.write(str(e) + "\n")
-        sys.stderr.write("Please install ElementTree module from\n")
-        sys.stderr.write("http://effbot.org/zone/element-index.htm\n")
-        sys.exit(1)
-
-try:
-    ## Remove 'MANIFEST' file to force
-    ## distutils to recreate it.
-    ## Only in "sdist" stage. Otherwise
-    ## it makes life difficult to packagers.
-    if sys.argv[1] == "sdist":
         os.unlink("MANIFEST")
-except:
-    pass
+    except OSError as e:
+        pass
 
 ## Re-create the manpage
 ## (Beware! Perl script on the loose!!)
-if sys.argv[1] == "sdist":
+if len(sys.argv) > 1 and sys.argv[1] == "sdist":
     if os.stat_result(os.stat("s3cmd.1")).st_mtime < os.stat_result(os.stat("s3cmd")).st_mtime:
         sys.stderr.write("Re-create man page first!\n")
         sys.stderr.write("Run: ./s3cmd --help | ./format-manpage.pl > s3cmd.1\n")
@@ -51,7 +53,7 @@ if not os.getenv("S3CMD_PACKAGING"):
     man_path = os.getenv("S3CMD_INSTPATH_MAN") or "share/man"
     doc_path = os.getenv("S3CMD_INSTPATH_DOC") or "share/doc/packages"
     data_files = [
-        (doc_path+"/s3cmd", [ "README.md", "INSTALL", "NEWS" ]),
+        (doc_path+"/s3cmd", [ "README.md", "INSTALL", "LICENSE", "NEWS" ]),
         (man_path+"/man1", [ "s3cmd.1" ] ),
     ]
 else:
@@ -69,7 +71,7 @@ setup(
     ## Packaging details
     author = "Michal Ludvig",
     author_email = "michal@logix.cz",
-    maintainer = "github.com/mdomsch, github.com/matteobar",
+    maintainer = "github.com/mdomsch, github.com/matteobar, github.com/fviard",
     maintainer_email = "s3tools-bugs@lists.sourceforge.net",
     url = S3.PkgInfo.url,
     license = S3.PkgInfo.license,
@@ -95,14 +97,19 @@ Authors:
         'Operating System :: Microsoft :: Windows',
         'Operating System :: POSIX',
         'Operating System :: Unix',
+        'Programming Language :: Python :: 2',
         'Programming Language :: Python :: 2.6',
         'Programming Language :: Python :: 2.7',
-        'Programming Language :: Python :: 2 :: Only',
+        'Programming Language :: Python :: 3',
+        'Programming Language :: Python :: 3.3',
+        'Programming Language :: Python :: 3.4',
+        'Programming Language :: Python :: 3.5',
+        'Programming Language :: Python :: 3.6',
         'Topic :: System :: Archiving',
         'Topic :: Utilities',
     ],
 
     install_requires = ["python-dateutil", "python-magic"]
-    )
+)
 
 # vim:et:ts=4:sts=4:ai
